requirements.txt
APScheduler==3.10.4
fake_useragent==1.5.1
Flask==3.1.0
PyYAML==5.4.1
Requests==2.32.3
urllib3==1.26.5
